Introduced the code to run the tests on grid environment.

+++ OpenCartV121/src/test/java/testBase/BaseTestClass.java
@@ -3,6 +3,7 @@ package testBase;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
+import java.net.URL;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -13,11 +14,14 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.safari.SafariDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -37,17 +41,39 @@ public class BaseTestClass {
 		FileReader file = new FileReader(".//src//test//resources//config.properties");
 		prop = new Properties();
 		prop.load(file);
-		
-		
-		
 		logger = LogManager.getLogger(this.getClass());
-		switch(browser.toLowerCase()) {
-		case "chrome" : driver = new ChromeDriver(); break;
-		case "edge" : driver = new EdgeDriver(); break;
-		case "firefox" : driver = new FirefoxDriver(); break;
-		case "safari" : driver = new SafariDriver(); break;
-		default: logger.info("Invalid Browser...."); return;
+		if(prop.getProperty("execution_env").equals("remote")) {
+			
+			DesiredCapabilities capabilities = new DesiredCapabilities();
+			switch(os.toLowerCase()) {
+			case "windows" : capabilities.setPlatform(Platform.WIN11); break;
+			case "mac" : capabilities.setPlatform(Platform.MAC); break;
+			case "linux": capabilities.setPlatform(Platform.LINUX); break;
+			default : logger.info("Invalid OS"); 
+			}
+			
+			switch(browser.toLowerCase()) {
+			case "chrome" : capabilities.setBrowserName("chrome"); break;
+			case "edge" :  capabilities.setBrowserName("MicrosoftEdge"); break;
+			case "firefox" :  capabilities.setBrowserName("firefox"); break;
+			case "safari" :  capabilities.setBrowserName("safari"); break;
+			default: logger.info("Invalid Browser...."); return;
+			}
+			
+			// add /wd/hub to the grid URL
+			driver = new RemoteWebDriver(new URL("http://localhost:4444/"), capabilities);
+			
 		}
+		if(prop.getProperty("execution_env").equals("local")) {
+			switch(browser.toLowerCase()) {
+			case "chrome" : driver = new ChromeDriver(); break;
+			case "edge" : driver = new EdgeDriver(); break;
+			case "firefox" : driver = new FirefoxDriver(); break;
+			case "safari" : driver = new SafariDriver(); break;
+			default: logger.info("Invalid Browser...."); return;
+			}
+		}
+		
 		
 		driver.manage().deleteAllCookies();
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
